requirements.txt
fastapi
uvicorn
pygithub
google-generativeai
python-dotenv
pydantic
httpx
google-adk
